Move event into subpackage and reduce method vis
